refactor: rename testimonials helper to utils
